+++ src/infrastructure.logging/src/transports/outputChannelTransport.ts
@@ -1,35 +1,42 @@
-// vscode references
-import * as VsCodeTypes from 'vscode';
+import { OutputChannel } from 'vscode';
+
+import { ILoggingOptions } from 'core.logging';
+
+import { ILoggerTransport } from './iLoggerTransport';
 
 const { Transport } = require('winston');
 
 const MESSAGE = Symbol.for('message');
 
-class OutputChannelTransport extends Transport {
+export class OutputChannelTransport extends Transport implements ILoggerTransport {
 
-  outputChannel: VsCodeTypes.OutputChannel;
-
-  constructor(outputChannel: VsCodeTypes.OutputChannel, transportOptions) {
-    super(transportOptions);
+  constructor(outputChannel: OutputChannel, logging: ILoggingOptions) {
+    super({ level: logging.level });
     this.outputChannel = outputChannel;
+    this.logging = logging;
+  }
+
+  outputChannel: OutputChannel;
+
+  logging: ILoggingOptions;
+
+  get name() {
+    return this.outputChannel.name;
   }
 
   log(entry, callback) {
 
     setImmediate(() => {
       this.emit('logged', entry)
-
       this.outputChannel.appendLine(`${entry[MESSAGE]}`);
     });
 
     callback();
   }
 
-}
+  updateLevel() {
+    this.logging.defrost();
+    super.level = this.logging.level;
+  }
 
-export function createOutputChannelTransport(
-  channel: VsCodeTypes.OutputChannel,
-  transportOptions
-) {
-  return new OutputChannelTransport(channel, transportOptions);
 }